LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/registered-property-revert-expected.txt


PASS Inherited registered custom property can be reverted 
PASS Non-inherited registered custom property can be reverted 
FAIL Non-inherited registered custom property can be reverted in animation assert_equals: expected "50px" but got "0px"
FAIL Inherited registered custom property can be reverted in animation assert_equals: expected "50px" but got "0px"
